Implemented CRUD operations with Events/profile Page.

+++ Assignment2/crud_app/assets/js/index.js
@@ -63,6 +63,31 @@ $("#add_user").submit(function(event){
 })
 
 
+$("#add_event").submit(function(event){
+	event.preventDefault();
+			var unindexed_array = $(this).serializeArray();
+			var data = {}
+
+			$.map(unindexed_array, function(n, i){
+					data[n['name']] = n['value']
+			})
+
+//		if (validateForm(data)) {  NEED TO VALIDATE LATER
+			var request = {
+					"url" : `http://localhost:3000/api/events`,
+					"method" : "POST",
+					"data" : data
+			}
+
+			$.ajax(request).done(function(response){
+					alert("Data Inserted Successfully!");
+			})
+
+//    }
+
+})
+
+
 $("#update_user").submit(function(event){
     event.preventDefault();
 
@@ -89,6 +114,32 @@ $("#update_user").submit(function(event){
 })
 
 
+$("#update_event").submit(function(event){
+    event.preventDefault();
+
+			var unindexed_array = $(this).serializeArray();
+			var data = {}
+
+			$.map(unindexed_array, function(n, i){
+					data[n['name']] = n['value']
+			})
+
+//		if (validateForm(data)) { NEED TO DO VALIDATION LATER
+			var request = {
+					"url" : `http://localhost:3000/api/event/${data.id}`,
+					"method" : "PUT",
+					"data" : data
+			}
+
+			$.ajax(request).done(function(response){
+					alert("Data Updated Successfully!");
+			})
+
+//    }
+
+})
+
+
 //Deletes a user and updates the page to show it
 if(window.location.pathname == "/"){
     $(".table tbody td a.delete").click(function(){
@@ -96,6 +147,26 @@ if(window.location.pathname == "/"){
 
         var request = {
             "url" : `http://localhost:3000/api/users/${id}`,
+            "method" : "DELETE"
+        }
+
+        if(confirm("Do you really want to delete this record?")){
+            $.ajax(request).done(function(response){
+                alert("Data Deleted Successfully!");
+                location.reload();
+            })
+        }
+
+    })
+}
+
+//Deletes a user and updates the page to show it
+if(window.location.pathname == "/profile"){
+    $(".table tbody td a.delete").click(function(){
+        var id = $(this).attr("data-id")
+
+        var request = {
+            "url" : `http://localhost:3000/api/events/${id}`,
             "method" : "DELETE"
         }
